remove contracts changes

+++ contracts/src/eip4788/interfaces/IBeaconVerifier.sol
@@ -60,12 +60,11 @@ interface IBeaconVerifier {
     /// given timestamp. Reverts if proof invalid.
     /// @param timestamp `uint64` timestamp of the parent beacon block.
     /// @param executionNumberProof `bytes32[]` proof of the execution number.
-    /// @param blockNumber `uint256` execution number of the parent beacon
-    /// block.
+    /// @param blockNumber `uint64` execution number of the parent beacon block.
     function verifyExecutionNumber(
         uint64 timestamp,
         bytes32[] calldata executionNumberProof,
-        uint256 blockNumber
+        uint64 blockNumber
     )
         external
         view;
